0817 - 타입스크립트로 전환

+++ washcar-newcar/src/pages/LoginRedirect.tsx
@@ -2,16 +2,22 @@ import { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
 const LoginRedirect = () => {
-  const { token } = useParams();
+  const { token } = useParams<{ token: string }>();
   const navigate = useNavigate();
-
+  
   useEffect(() => {
+    if (!token) {
+      navigate('/');
+      return;
+    }
+
     localStorage.clear();
     localStorage.setItem('jwt', token);
     navigate('/');
+    return;
   }, []);
 
-  return <></>;
+  return null;
 };
 
 export default LoginRedirect;
